User email and validation for scheduling deletion
